+++ src/main/kotlin/it/czerwinski/intellij/wavefront/lang/psi/util/TextureFiles.kt
@@ -18,12 +18,14 @@ package it.czerwinski.intellij.wavefront.lang.psi.util
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.openapi.vfs.findFile
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiManager
 import com.intellij.psi.search.FilenameIndex
 import com.intellij.psi.search.GlobalSearchScope
 import com.jogamp.opengl.util.texture.TextureIO
+import java.io.File
 
 private val textureFilesExtensions = listOf(
     TextureIO.DDS,
@@ -45,12 +47,22 @@ fun Project.findAllTextureFiles(): List<PsiFile> =
         PsiManager.getInstance(this).findFile(virtualFile)
     }
 
-fun Project.findMatchingTextureFiles(filename: String): List<PsiFile> =
-    findMatchingTextureVirtualFiles(filename)
+fun Project.findMatchingTextureFiles(filename: String, relativeTo: VirtualFile? = null): List<PsiFile> =
+    findMatchingTextureVirtualFiles(filename, relativeTo)
         .mapNotNull { virtualFile -> PsiManager.getInstance(this).findFile(virtualFile) }
 
-fun Project.findMatchingTextureVirtualFiles(filename: String): Collection<VirtualFile> =
-    FilenameIndex.getVirtualFilesByName(filename, GlobalSearchScope.allScope(this))
+fun Project.findMatchingTextureVirtualFiles(
+    filename: String,
+    relativeTo: VirtualFile? = null
+): Collection<VirtualFile> {
+    val baseDirectory = if (relativeTo?.isDirectory != false) relativeTo else relativeTo.parent
+    val relativeFile = baseDirectory?.findFile(relativePath = filename)
+    if (relativeFile != null) {
+        return listOf(relativeFile)
+    }
+    val rawFilename = File(filename).name
+    return FilenameIndex.getVirtualFilesByName(rawFilename, GlobalSearchScope.allScope(this))
+}
 
 fun PsiElement?.isTextureFile(): Boolean =
     this is PsiFile && textureFilesExtensions.any { ext -> this.name.endsWith(".$ext", ignoreCase = true) }